+++ src/main/java/org/example/services/ArticleService.java
@@ -1,6 +1,8 @@
 package org.example.services;
 
 import org.example.entities.ArticlesEntity;
+import org.example.exceptions.ArticleExceptions.ArticleExistedException;
+import org.example.exceptions.ArticleExceptions.ArticleNotExistException;
 import org.example.repositories.ArticleRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -13,19 +15,37 @@ public class ArticleService {
     @Autowired()
     private ArticleRepository articleRepository;
 
-    public Optional<ArticlesEntity> getArticleById(UUID articleId) {
-        return articleRepository.findById(articleId);
+    public ArticlesEntity getArticleById(UUID articleId) {
+        return articleRepository.findById(articleId)
+                .orElseThrow(() -> new ArticleNotExistException("Article '" + articleId + "' not exist"));
     }
 
-    public Optional<ArticlesEntity> getArticleByName(String articleName) {
-        return articleRepository.getArticleByName(articleName);
+    public ArticlesEntity getArticleByName(String articleName) {
+        return articleRepository.getArticleByName(articleName)
+                .orElseThrow(() -> new ArticleNotExistException("Article '" + articleName + "' not exist"));
     }
 
-    public ArticlesEntity saveArticle(ArticlesEntity articlesEntity) {
+    public ArticlesEntity createArticle(ArticlesEntity articlesEntity) {
+        if(articleRepository.existsByName(articlesEntity.getArticleName())) {
+            throw new ArticleExistedException("Article '" + articlesEntity.getArticleName() + "' existed");
+        }
+
         return articleRepository.save(articlesEntity);
     }
 
+    public ArticlesEntity updateArticle(ArticlesEntity articlesEntity) {
+        if(articleRepository.existsById(articlesEntity.getArticleId())) {
+            return articleRepository.save(articlesEntity);
+        }
+
+        throw new ArticleNotExistException("Article '" + articlesEntity.getArticleName() + "' not exist");
+    }
+
     public void deleteArticle(UUID articleId) {
-        articleRepository.deleteById(articleId);
+        if(articleRepository.existsById(articleId)) {
+            articleRepository.deleteById(articleId);
+        }
+
+        throw new ArticleNotExistException("Article '" + articleId + "' not exist");
     }
 }
