+++ src/main/java/org/example/services/ImagesOfAlbumService.java
@@ -1,6 +1,7 @@
 package org.example.services;
 
 import org.example.entities.ImagesOfAlbumEntity;
+import org.example.exceptions.ImageOfAlbumExceptions.ImageOfAlbumNotExistException;
 import org.example.repositories.ImagesOfAlbumRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -13,15 +14,28 @@ public class ImagesOfAlbumService {
     @Autowired()
     private ImagesOfAlbumRepository imagesOfAlbumRepository;
 
-    public Optional<ImagesOfAlbumEntity> getImageById(UUID imageId) {
-        return imagesOfAlbumRepository.findById(imageId);
+    public ImagesOfAlbumEntity getImageById(UUID imageId) {
+        return imagesOfAlbumRepository.findById(imageId)
+                .orElseThrow(() -> new ImageOfAlbumNotExistException("image '" + imageId + "' not exist"));
     }
 
-    public ImagesOfAlbumEntity saveImage(ImagesOfAlbumEntity imagesOfAlbumEntity) {
+    public ImagesOfAlbumEntity createImage(ImagesOfAlbumEntity imagesOfAlbumEntity) {
         return imagesOfAlbumRepository.save(imagesOfAlbumEntity);
     }
 
+    public ImagesOfAlbumEntity updateImage(ImagesOfAlbumEntity imagesOfAlbumEntity) {
+        if (imagesOfAlbumRepository.existsById(imagesOfAlbumEntity.getImageId())){
+            return imagesOfAlbumRepository.save(imagesOfAlbumEntity);
+        }
+
+        throw new ImageOfAlbumNotExistException("image '" + imagesOfAlbumEntity.getImageId() + "' not exist");
+    }
+
     public void deleteImage(UUID imageId) {
-        imagesOfAlbumRepository.deleteById(imageId);
+        if(imagesOfAlbumRepository.existsById(imageId)) {
+            imagesOfAlbumRepository.deleteById(imageId);
+        }
+
+        throw new ImageOfAlbumNotExistException("image '" + imageId + "' not exist");
     }
 }
